Jetpack Onboarding: Move Credentials State to jetpack/onboarding
