+++ client/state/jetpack-onboarding/reducer.js
@@ -9,30 +9,15 @@ import { mapValues, merge } from 'lodash';
  * Internal dependencies
  */
 import { createReducer, combineReducers, keyedReducer } from 'state/utils';
-import { jetpackOnboardingCredentialsSchema, jetpackSettingsSchema } from './schema';
+import { jetpackSettingsSchema } from './schema';
 import { normalizeSettings } from 'state/jetpack/settings/utils';
 import {
 	JETPACK_MODULE_ACTIVATE_SUCCESS,
 	JETPACK_MODULE_DEACTIVATE_SUCCESS,
 	JETPACK_MODULES_RECEIVE,
-	JETPACK_CONNECT_AUTHORIZE_RECEIVE,
-	JETPACK_ONBOARDING_CREDENTIALS_RECEIVE,
 	JETPACK_ONBOARDING_SETTINGS_SAVE_SUCCESS,
 	JETPACK_ONBOARDING_SETTINGS_UPDATE,
 } from 'state/action-types';
-
-export const credentialsReducer = keyedReducer(
-	'siteId',
-	createReducer(
-		{},
-		{
-			[ JETPACK_ONBOARDING_CREDENTIALS_RECEIVE ]: ( state, { credentials } ) => credentials,
-			[ JETPACK_CONNECT_AUTHORIZE_RECEIVE ]: () => undefined,
-		},
-		jetpackOnboardingCredentialsSchema
-	)
-);
-credentialsReducer.hasCustomPersistence = true;
 
 export const settingsReducer = keyedReducer(
 	'siteId',
@@ -82,6 +67,5 @@ export const settingsReducer = keyedReducer(
 settingsReducer.hasCustomPersistence = true;
 
 export default combineReducers( {
-	credentials: credentialsReducer,
 	settings: settingsReducer,
 } );
