+++ client/state/jetpack-onboarding/schema.js
@@ -1,16 +1,5 @@
 /** @format */
 
-export const jetpackOnboardingCredentialsSchema = {
-	type: 'object',
-	additionalProperties: false,
-	required: [ 'siteUrl', 'token', 'userEmail' ],
-	properties: {
-		siteUrl: { type: 'string' },
-		token: { type: 'string' },
-		userEmail: { type: 'string' },
-	},
-};
-
 export const jetpackOnboardingSettingsSchema = {
 	type: 'object',
 	additionalProperties: false,
