+++ client/state/jetpack-onboarding/test/actions.js
@@ -4,7 +4,6 @@
  * Internal dependencies
  */
 import {
-	receiveJetpackOnboardingCredentials,
 	regeneratePostByEmail,
 	requestJetpackSettings,
 	saveJetpackSettings,
@@ -12,7 +11,6 @@ import {
 	updateJetpackSettings,
 } from '../actions';
 import {
-	JETPACK_ONBOARDING_CREDENTIALS_RECEIVE,
 	JETPACK_ONBOARDING_SETTINGS_REQUEST,
 	JETPACK_ONBOARDING_SETTINGS_SAVE,
 	JETPACK_ONBOARDING_SETTINGS_SAVE_SUCCESS,
@@ -20,24 +18,6 @@ import {
 } from 'state/action-types';
 
 describe( 'actions', () => {
-	describe( 'receiveJetpackOnboardingCredentials()', () => {
-		test( 'should return a jetpack onboarding credentials receive action object', () => {
-			const credentials = {
-				token: 'abcd1234',
-				siteUrl: 'http://yourgroovydomain.com/',
-				userEmail: 'somebody@yourgroovydomain.com',
-			};
-			const siteId = 12345678;
-			const action = receiveJetpackOnboardingCredentials( siteId, credentials );
-
-			expect( action ).toEqual( {
-				type: JETPACK_ONBOARDING_CREDENTIALS_RECEIVE,
-				siteId,
-				credentials,
-			} );
-		} );
-	} );
-
 	describe( 'requestJetpackSettings()', () => {
 		test( 'should return a jetpack settings request action object', () => {
 			const siteId = 12345678;
