+++ client/state/jetpack/onboarding/actions.js
@@ -1,0 +1,12 @@
+/** @format */
+
+/**
+ * Internal dependencies
+ */
+import { JETPACK_ONBOARDING_CREDENTIALS_RECEIVE } from 'state/action-types';
+
+export const receiveJetpackOnboardingCredentials = ( siteId, credentials ) => ( {
+	type: JETPACK_ONBOARDING_CREDENTIALS_RECEIVE,
+	siteId,
+	credentials,
+} );
