+++ client/state/jetpack/onboarding/schema.js
@@ -1,0 +1,12 @@
+/** @format */
+
+export const jetpackOnboardingCredentialsSchema = {
+	type: 'object',
+	additionalProperties: false,
+	required: [ 'siteUrl', 'token', 'userEmail' ],
+	properties: {
+		siteUrl: { type: 'string' },
+		token: { type: 'string' },
+		userEmail: { type: 'string' },
+	},
+};
