+++ client/state/jetpack/onboarding/test/actions.js
@@ -1,0 +1,27 @@
+/** @format */
+
+/**
+ * Internal dependencies
+ */
+import { receiveJetpackOnboardingCredentials } from '../actions';
+import { JETPACK_ONBOARDING_CREDENTIALS_RECEIVE } from 'state/action-types';
+
+describe( 'actions', () => {
+	describe( 'receiveJetpackOnboardingCredentials()', () => {
+		test( 'should return a jetpack onboarding credentials receive action object', () => {
+			const credentials = {
+				token: 'abcd1234',
+				siteUrl: 'http://yourgroovydomain.com/',
+				userEmail: 'somebody@yourgroovydomain.com',
+			};
+			const siteId = 12345678;
+			const action = receiveJetpackOnboardingCredentials( siteId, credentials );
+
+			expect( action ).toEqual( {
+				type: JETPACK_ONBOARDING_CREDENTIALS_RECEIVE,
+				siteId,
+				credentials,
+			} );
+		} );
+	} );
+} );
